Revision of Towards Expertise/ Practice advice
A coworker recommended adding a link to HackerRank. (#28) Since this was yet another "challenges" resource ... but a unique one (the challenges are smaller/ bite-size) ... seemed useful, but also imbalanced the "Towards Expertise" section towards challenges/competitions, a little too much.

So that made me want to revise further, and emphasize that the challenges/competition way offers practice -- but that it should always be paired with understanding ... and also that the "independent research" approach -- doing practice projects driven by your own interests -- is complementary.

Slipped in a link to Data is Plural, as I've been wanting to (#30)