TensorFlow note -> TensorFlow, Theano, Torch note
Addresses #27: TensorFlow is exciting but it's not the only thing of its kind. Also needed better context. Also wanted to link to some other good resources like a comparison of TensorFlow and similar frameworks (that came first) -- Theano, Torch. And tutorials.

I think I've got that all in, pretty concisely