Update the section about places to share work (#168)
* Update the section about places to share work 

I meant to link to OpenReview.net a while ago, and mostly I just kept meaning to find whether I should link to another resource that's "equivalent" to that project somehow, but maybe a different organization, just to try and be complete, or just in case I am unaware of some reason it's "narrow" to link to that organization while forgetting another one. Or, to put that differently: if anybody wants to submit a PR adding another organization/platform to the list of Peer Review related links, please do so! At any rate, I *WANT* to link to OpenReview.net, and I didn't mean to wait so long to do so.