Update README.md
Removed a bit of description on an item, but left the entry on it. It's not due to anything wrong, I just had occasion to look back at the guide, and it was kind of distracting to have the blurb about performance metrics, in-line with so much other info that does _not_ get into the weeds about [those same] performance metrics. So it's an edit for the sake of the guide overall. I don't think anybody will mind.