Revise a couple of links at the end
There were some cases where a given name was among "the" names you often heard listed together, and so it was contextually relevant to include more of that info, in the links in the guide here. In some of those cases, there are now related projects that are quite relevant, and it would be strange to include some of the random links without updating to include alternatives, similar software, more recent projects, etc. So, there's nothing wrong with the links I'm removing here; they're just well-known projects, and they're linked prominently from the other links in the guide about NLP: awesome-nlp, etc.

The reason I had a little more specific links about NLP in the first place, is just because it's been a subject of interest in the past (going back to college actually, with NLP). But since I'm not working in NLP actively, I thought maybe it's weird to keep this highlight of SpaCy just because I personally had some good experiences; it's imbalanced in the context of the other recommendations, and the other projects in the past few years. Anyways the guide still links to explosion/thinc elsewhere too, and again the point is that awesome-nlp, etc, link to SpaCy; I'm trying to make the guide generic/agnostic about NLP since it's more niche and specific than this particular guide.