FAQ += re: package managers (pragmatic tips, while you're learning ML)
This needs additional revision! But I recently answered this for someone and I wanted to jot something based on that, here. A lot of folks are still unsure which thing they should use, and I've opted to link to other good blog posts, but maybe I can put a paragraph or something, to give a fast answer.

But yeah it needs some revision. (As-is, it might not be worth merging.)