Remove aside about podcasts, section got stale
It was in here because I found it beneficial learning earlier on when info was harder to find and harder to keep up with, in this field. Now, it feels too in the weeds and like it's picking one podcast versus others, when really I just don't have time to do a roundup of podcasts on ML topics.

Anyways, it doesn't make sense to yak-shave when it was an aside in the first place.

In deleting this aside, it also deletes a couple extra links, but I actually JUST merged them recently,  as a matter of "checking items off a todo list" -- the links were gathered earlier and they were sitting in Issues. After deleting the podcast section, I reviewed the guide to see if these items still fit and I'm not sure where they'd fit. Though of course if you happen to see this diff and you wanna make a pull request, feel free to do so!